requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bleach==6.1.0
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
Django==4.2.6
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-crispy-forms==2.3
django-summernote==0.8.20.0
django-taggit==4.0.0
fontawesomefree==6.6.0
Pillow==10.0.1
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
tzdata==2023.3
webencodings==0.5.1
